+++ src/components/discussion/box/AuthorContentBtn.tsx
@@ -1,7 +1,13 @@
 import React from "react";
+import { useRouter } from "next/navigation";
+import { useRecoilValue } from "recoil";
 
 import Button from "@/components/designSystem/Button";
 import useModalController from "@/hooks/useModalController";
+import ConfirmModal from "@/components/modal/common/ConfirmModal";
+import { discussionAPI } from "@/api/discussionAPI";
+import { discussionAtom } from "@/constants/global/atoms";
+import { isLoggedInAtom } from "@/hooks/useUserAuth";
 
 import DeleteDiscussionBtn from "../btn/DeleteDiscussionBtn";
 import VerifyDiscussionBtn from "../btn/verifyDiscussionBtn";
@@ -10,9 +16,24 @@ interface props {
   isAuthor: boolean;
 }
 const AuthorContentBtn: React.FC<props> = ({ isAuthor }) => {
-  const { openModal } = useModalController();
+  const { discussionId } = useRecoilValue(discussionAtom);
+  const isLoggedIn = useRecoilValue(isLoggedInAtom);
+  const { modal, openModal, closeModal } = useModalController();
+  const router = useRouter();
   const discussionExitHandler = () => {
+    if (!isLoggedIn) {
+      return router.back();
+    }
     openModal("discussionExit");
+  };
+
+  const okconfirmHanldler = async () => {
+    const result = await discussionAPI.leftMyDiscussion(discussionId);
+    if (result.status === 200) router.back();
+    else {
+      alert("토의 나가기에 실패했습니다.");
+      closeModal("discussionExit");
+    }
   };
 
   return (
@@ -37,6 +58,14 @@ const AuthorContentBtn: React.FC<props> = ({ isAuthor }) => {
         >
           토의 나가기
         </Button>
+      )}
+      {modal.discussionExit && (
+        <ConfirmModal
+          title="이 토의에 대한 알림을 받으시겠습니까?"
+          content="토의의 끝나는 내용을 공유 받으실 수 있습니다."
+          onNegativeClick={() => router.back()}
+          onPositiveClick={okconfirmHanldler}
+        />
       )}
     </>
   );
